\name{racial}
\alias{racial}
\docType{data}

\title{Racial data}

\description{Proportion of white student enrollment in 56 school districts in Nassau County (Long Island, New York), for the 1992-1993 school year.}

\usage{data(racial)}

\format{A data frame with the following variables:

\describe{
\item{District}{School district.}
\item{PropWhite}{Proportion of white student enrolled.}
}
}

\source{Simonoff, S.J. (1996) Smoothing Methods in Statistics, Springer-Verlag, New York, p. 52}

\keyword{datasets}
