\name{suicide}
\alias{suicide}
\docType{data}

\title{Suicide data}

\description{Lengths of treatment spells (in days) of control patients in suicide study.}

\usage{data(suicide)}

\format{A vector of containing the lengths (days) of 86 spells of psychiatric treatment undergone by patients used as controls in a study of suicide risks.}

\source{Silverman, B. W. (1986) Density Estimation, Chapman & Hall, Tab 2.1.}

\keyword{datasets}
